refactor: object types & resolve fields
