+++ ecp-nest-backend/src/product-promotions/domain/object-types/product-promotion.type.ts
@@ -4,6 +4,13 @@ import { PromotionType } from 'src/promotions/domain/object-types/promotion.type
 
 @ObjectType()
 export class ProductPromotionType {
+  @Field(() => Date)
+  createdAt: Date;
+
+  @Field(() => Date, { nullable: true })
+  updatedAt?: Date;
+
+  // Relations
   @Field(() => ProductType)
   product: ProductType;
