+++ ecp-nest-backend/src/product-tags/domain/object-types/product-tag.type.ts
@@ -4,6 +4,13 @@ import { TagType } from 'src/tags/domain/object-types/tag.type';
 
 @ObjectType()
 export class ProductTagType {
+  @Field(() => Date)
+  createdAt: Date;
+
+  @Field(() => Date, { nullable: true })
+  updatedAt?: Date;
+
+  // Relations
   @Field(() => ProductType)
   product: ProductType;
