+++ ecp-nest-backend/src/seasons/domain/object-types/season.type.ts
@@ -16,6 +16,12 @@ export class SeasonType {
   @Field(() => Date, { nullable: true })
   endDate?: Date;
 
+  @Field(() => Date)
+  createdAt: Date;
+
+  @Field(() => Date, { nullable: true })
+  updatedAt?: Date;
+
   // Relations
   @Field(() => CategoryType)
   categories: CategoryType[];
