+++ ecp-nest-backend/src/shop-order-locations/domain/dtos/graphql/inputs/shop-order-location.input.ts
@@ -1,5 +1,5 @@
-import { Field, ID, InputType } from '@nestjs/graphql';
-import { PartialType } from '@nestjs/swagger';
+import { Field, ID, InputType, PartialType } from '@nestjs/graphql';
+
 import { IsNotEmpty, IsUUID } from 'class-validator';
 
 @InputType()
