+++ ecp-nest-backend/src/shopping-cart-product-items/domain/dtos/graphql/inputs/shopping-cart-product-item.input.ts
@@ -1,5 +1,4 @@
-import { Field, ID, InputType, Int } from '@nestjs/graphql';
-import { PartialType } from '@nestjs/swagger';
+import { Field, ID, InputType, Int, PartialType } from '@nestjs/graphql';
 import { IsInt, IsNotEmpty, IsPositive, IsUUID, Min } from 'class-validator';
 
 @InputType()
