+++ ecp-nest-backend/src/shopping-cart-product-items/domain/object-types/shopping-cart-product-item.type.ts
@@ -10,6 +10,12 @@ export class ShoppingCartProductItemType {
   @Field(() => Float)
   quantity: number;
 
+  @Field(() => Date)
+  createdAt: Date;
+
+  @Field(() => Date, { nullable: true })
+  updatedAt?: Date;
+
   // Relations
   @Field(() => ShoppingCartType)
   shoppingCart: ShoppingCartType;
