+++ ecp-nest-backend/src/shopping-carts/domain/object-types/shopping-cart.type.ts
@@ -6,6 +6,12 @@ export class ShoppingCartType {
   @Field(() => ID)
   id: string;
 
+  @Field(() => Date)
+  createdAt: Date;
+
+  @Field(() => Date, { nullable: true })
+  updatedAt?: Date;
+
   // Relations
   @Field(() => UserObjType)
   user: UserObjType;
