+++ ecp-nest-backend/src/stores/interface-adapters/resolvers/store.resolver.ts
@@ -1,5 +1,6 @@
 import { ParseUUIDPipe } from '@nestjs/common';
-import { Args, Resolver, Query, Mutation } from '@nestjs/graphql';
+import { Args, Resolver, Query, Mutation, ResolveField } from '@nestjs/graphql';
+import { BillboardType } from 'src/billboard/domain/object-types/billboard.type';
 import {
   PaginationArgs,
   SearchArgs,
@@ -10,11 +11,18 @@ import {
   UpdateStoreInput,
 } from 'src/stores/domain/dtos/graphql/inputs/store.input';
 import { IStore } from 'src/stores/domain/entities/store.entity';
-import { StoreType } from 'src/stores/domain/object-types/store-type';
+import { StoreType } from 'src/stores/domain/object-types/store.type';
+import { BillboardUseCases } from '../../../billboard/application/use-cases/billboard-use-cases';
+import { CategoryType } from 'src/categories/domain/object-types/category.type';
+import { CategoryUseCases } from 'src/categories/application/use-cases/category-use-cases';
 
 @Resolver(() => StoreType)
 export class StoreResolver {
-  constructor(private storeUseCases: StoreUseCases) {}
+  constructor(
+    private storeUseCases: StoreUseCases,
+    private billboardUseCases: BillboardUseCases,
+    private categoryUseCases: CategoryUseCases,
+  ) {}
 
   @Query(() => [StoreType], { name: 'stores' })
   async getAllStores(
@@ -56,5 +64,28 @@ export class StoreResolver {
     @Args('id', { type: () => String }, ParseUUIDPipe) id: string,
   ) {
     return this.storeUseCases.removeStore(id);
+  }
+
+  // Resolve fields
+  @ResolveField(() => [BillboardType], { name: 'billboards' })
+  async getBillboards(
+    @Args() paginationArgs: PaginationArgs,
+    @Args() searchArgs: SearchArgs,
+  ) {
+    return this.billboardUseCases.getAllBillboards({
+      paginationArgs,
+      searchArgs,
+    });
+  }
+
+  @ResolveField(() => [CategoryType], { name: 'categories' })
+  async getCategories(
+    @Args() paginationArgs: PaginationArgs,
+    @Args() searchArgs: SearchArgs,
+  ) {
+    return this.categoryUseCases.getAllCategories({
+      paginationArgs,
+      searchArgs,
+    });
   }
 }
