+++ ecp-nest-backend/src/user-addresses/domain/dtos/graphql/inputs/user-address.input.ts
@@ -1,5 +1,4 @@
-import { Field, ID, InputType } from '@nestjs/graphql';
-import { PartialType } from '@nestjs/swagger';
+import { Field, ID, InputType, PartialType } from '@nestjs/graphql';
 import { IsBoolean, IsNotEmpty, IsUUID } from 'class-validator';
 
 @InputType()
