+++ ecp-nest-backend/src/user-addresses/domain/object-types/user-address.type.ts
@@ -7,6 +7,12 @@ export class UserAddressType {
   @Field(() => Boolean, { nullable: true })
   isDefault?: boolean;
 
+  @Field(() => Date)
+  createdAt: Date;
+
+  @Field(() => Date, { nullable: true })
+  updatedAt?: Date;
+
   // Relations
   @Field(() => UserObjType)
   user: UserObjType;
