+++ ecp-nest-backend/src/user-payment-methods/domain/dtos/graphql/inputs/user-payment-method.input.ts
@@ -1,5 +1,5 @@
-import { Field, ID, InputType } from '@nestjs/graphql';
-import { PartialType } from '@nestjs/swagger';
+import { Field, ID, InputType, PartialType } from '@nestjs/graphql';
+
 import {
   IsBoolean,
   IsCreditCard,
