+++ ecp-nest-backend/src/variation-options/domain/object-types/variation-option.type.ts
@@ -1,6 +1,6 @@
 import { ObjectType, Field, ID } from '@nestjs/graphql';
 import { ProductConfigurationType } from 'src/product-configurations/domain/object-types/product-configuration.type';
-import { VariationType } from 'src/variations/domain/object-types/variation.entity.type';
+import { VariationType } from 'src/variations/domain/object-types/variation.type';
 
 @ObjectType()
 export class VariationOptionType {
@@ -9,6 +9,12 @@ export class VariationOptionType {
 
   @Field(() => String)
   value: string;
+
+  @Field(() => Date)
+  createdAt: Date;
+
+  @Field(() => Date, { nullable: true })
+  updatedAt?: Date;
 
   // Relations
   @Field(() => [ProductConfigurationType])
